+++ deepdiff/deepbase.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from __future__ import print_function
+import difflib
+
+class DeepBase:
+    """Base class for deep-diff enabled containers"""
+    pass
