+++ test/test_deepdiff.py
@@ -10,13 +10,10 @@ python -m unittest tests.DeepDiffTestCase.test_list_of_sets_difference_ignore_or
 """
 import unittest
 from decimal import Decimal
-from sys import version
-from copy import copy
 
 from deepdiff.deepdiff import DeepDiff
 
-from . import CustomClass
-from . import py3
+from .setup import *
 
 
 class DeepDiffTestCase(unittest.TestCase):
