+++ test/test_deepset.py
@@ -9,14 +9,11 @@ To run a specific test, run this from the root of repo:
 python -m unittest tests.DeepDiffTestCase.test_list_of_sets_difference_ignore_order
 """
 import unittest
-from decimal import Decimal
-from sys import version
 from copy import copy
 
 from deepdiff.deepset import DeepSet
 
-from . import CustomClass
-from . import py3
+from .setup import *
 
 
 class DeepSetTestCase(unittest.TestCase):
@@ -53,6 +50,8 @@ class DeepSetTestCase(unittest.TestCase):
 
 
 
+    # x in s
+
     def test_custom_object_in_deepset(self):
         member1 = CustomClass(13, 37)
         member2 = CustomClass(13, 37)
@@ -66,6 +65,8 @@ class DeepSetTestCase(unittest.TestCase):
         self.assertTrue(member2 not in t1)
 
 
+
+    # isdisjoint()
 
     def test_disjoint_basic_true(self):
         t1 = DeepSet({"vegan", "for", "life"})
@@ -113,6 +114,8 @@ class DeepSetTestCase(unittest.TestCase):
         self.assertEqual(t2 & t3, {member2})
 
 
+
+    # difference(), -
 
     def test_sub_trivial(self):
         fibonacci = DeepSet({1, 2, 3, 5, 8, 13})
